Fix linters errors about cycle of files
- Added 2 new files for better practice:

1- books-constructor.js
2- render
