+++ index.js
@@ -1,9 +1,13 @@
-/* eslint-disable max-classes-per-file */
+/* eslint-disable no-unused-vars */
+// Added disable no-unused-vars because this ones are called only for the events of the nav bar
 
-import { bookList } from './modules/books.js';
+import { bookList, renderBook } from './modules/books.js';
 import { luxon } from './modules/luxon.js';
-import { renderBook, showTitle, storage } from './modules/functions.js';
-import { navNewBook, navMyList, navHome, navNavLink } from './modules/events.js'
+import showTitle from './modules/functions.js';
+import storage from './modules/storage.js';
+import {
+  navNewBook, navMyList, navHome, navNavLink,
+} from './modules/events.js';
 
 const time = document.getElementById('dateTime');
 time.innerHTML = luxon.DateTime.now().toLocaleString(
@@ -11,7 +15,7 @@ time.innerHTML = luxon.DateTime.now().toLocaleString(
 );
 
 const addBookButton = document.getElementById('add-book');
-let getMyLibrary = storage();
+const getMyLibrary = storage();
 
 addBookButton.addEventListener('click', () => {
   bookList.addBook();
@@ -19,5 +23,5 @@ addBookButton.addEventListener('click', () => {
 
 window.addEventListener('load', () => {
   renderBook(getMyLibrary);
-  showTitle(getMyLibrary)
+  showTitle(getMyLibrary);
 });
