+++ modules/books-constructor.js
@@ -1,0 +1,6 @@
+export default class Books {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
