+++ modules/events.js
@@ -46,5 +46,6 @@ const navNavLink = contactNavLink.addEventListener('click', (e) => {
   myContact.classList.remove('d-none');
 });
 
-
-export { navNewBook, navMyList, navHome, navNavLink };
+export {
+  navNewBook, navMyList, navHome, navNavLink,
+};
